requirements.txt
fastapi
uvicorn
scikit-learn
numpy
joblib
pandas
tensorflow
pillow
python-multipart
requests
